pyproject.toml
[project]
name = "agentic-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "langchain[google-genai]>=1.2.10",
    "langchain-google-genai>=4.2.0",
    "langchain-openai>=1.1.10",
    "openai-agents>=0.9.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.41.0",
    "langchain-groq>=1.1.2",
]
